+++ test/test_unit.jl
@@ -7,6 +7,8 @@ using LinearAlgebra: norm, dot
 using SparseArrays
 using DelimitedFiles: readdlm
 
+using ForwardDiff
+
 # Use Convex and ECOS to load the extension that extends functions for testing
 # PERK Single p2 Constructors
 using Convex: Convex
@@ -281,6 +283,20 @@ end
 
         @test Trixi.gauss_nodes_weights(3)[1] ≈ [-sqrt(3 / 5), 0.0, sqrt(3 / 5)]
         @test Trixi.gauss_nodes_weights(3)[2] ≈ [5 / 9, 8 / 9, 5 / 9]
+    end
+
+    @testset "boundary interpolation" begin
+        for p in 1:7
+            basis = LobattoLegendreBasis(p)
+            nodes = basis.nodes
+            weights = basis.weights
+
+            Lhat_minus1 = Trixi.calc_Lhat(-1.0, nodes, weights)
+            @test basis.inverse_weights[1] == Lhat_minus1[1]
+
+            Lhat_plus1 = Trixi.calc_Lhat(1.0, nodes, weights)
+            @test basis.inverse_weights[p + 1] == Lhat_plus1[p + 1]
+        end
     end
 
     @testset "multiply_dimensionwise" begin
@@ -762,6 +778,31 @@ end
                                                                   surface_fluxes,
                                                                   equations)))
     end
+end
+
+@timed_testset "Test consistency (fluxes, entropy/cons2entropy) for NonIdealCompressibleEulerEquations1D" begin
+    eos = VanDerWaals(; a = 10, b = 0.01, R = 287, gamma = 1.4)
+    equations = NonIdealCompressibleEulerEquations1D(eos)
+    q = SVector(2.0, 0.1, 10.0)
+    V, v1, T = q
+    u = prim2cons(q, equations)
+
+    @test density(u, equations) ≈ 0.5
+    @test velocity(u, equations) ≈ 0.1
+    @test density_pressure(u, equations) ≈ u[1] * pressure(V, T, eos)
+    @test energy_internal(u, equations) ≈ energy_internal(V, T, eos)
+
+    @test ForwardDiff.gradient(u -> entropy(u, equations), u) ≈
+          cons2entropy(u, equations)
+    @test flux_lax_friedrichs(u, u, 1, equations) ≈ flux(u, 1, equations)
+    @test flux_hll(u, u, 1, equations) ≈ flux(u, 1, equations)
+
+    # check that the fallback temperature and specialized temperature 
+    # return the same value 
+    V, v1, T = cons2prim(u, equations)
+    e = energy_internal(V, T, eos)
+    @test temperature(V, e, eos) ≈
+          invoke(temperature, Tuple{Any, Any, Trixi.AbstractEquationOfState}, V, e, eos)
 end
 
 @timed_testset "StepsizeCallback" begin